feat: move out components to components folder
