+++ pages/alpha/courses/web-development.tsx
@@ -1,7 +1,10 @@
 import Head from "next/head";
 import Layout from "../../../components/Layout/Layout";
-import { BookOpenIcon, CheckCircleIcon } from "@heroicons/react/outline";
-import clsx from "clsx";
+import {
+  CourseContens,
+  type IContent,
+} from "../../../components/Course/CourseContents";
+import { CircularProgressBar } from "../../../components/Course/CircularProgressBar";
 
 // ! This is hardcoded and needs to be dynamic
 // ! I assume the route also should like [course-name].tsx , not web-development.tsx. Keep in mind it is hardcoded for now!
@@ -59,82 +62,6 @@ const WebDevelopmentCourse = () => {
         </div>
       </Layout>
     </>
-  );
-};
-
-interface CourseContensProps {
-  title: string;
-  contents: IContent[];
-}
-
-interface IContent {
-  title: string;
-  completed?: boolean;
-}
-
-const CourseContens = ({ title, contents }: CourseContensProps) => {
-  return (
-    <div className="bg-neutral-900">
-      <div className="p-4 border-b border-neutral-500">
-        <p className="text-lg font-semibold leading-none">{title}</p>
-      </div>
-
-      <div className="p-4">
-        <ul className="flex flex-col gap-4">
-          {contents.map((c) => (
-            <li key={c.title} className="flex items-center justify-between">
-              <div className="flex items-center gap-3">
-                <BookOpenIcon className="w-6" />
-                <p className="font-medium">{c.title}</p>
-              </div>
-              <CheckCircleIcon
-                className={clsx("w-6", {
-                  "text-pink-600": c.completed,
-                })}
-              />
-            </li>
-          ))}
-        </ul>
-      </div>
-    </div>
-  );
-};
-
-const CircularProgressBar = ({ progress }: { progress: number }) => {
-  const strokeWidth = 10; // Adjust this value to change the thickness of the outline
-  const radius = 50 - strokeWidth / 2;
-  const circumference = 2 * Math.PI * radius;
-
-  return (
-    <div className="flex items-center justify-center overflow-hidden ">
-      <svg
-        className="w-32 h-32 transform translate-x-1 translate-y-1"
-        aria-hidden="true"
-      >
-        <circle
-          className="text-gray-300"
-          stroke-width="10"
-          stroke="currentColor"
-          fill="transparent"
-          r="50"
-          cx="60"
-          cy="60"
-        />
-        <circle
-          className="text-pink-600"
-          stroke-width="10"
-          strokeDasharray={circumference}
-          strokeDashoffset={circumference - (progress / 100) * circumference}
-          stroke-linecap="round"
-          stroke="currentColor"
-          fill="transparent"
-          r="50"
-          cx="60"
-          cy="60"
-        />
-      </svg>
-      <span className="absolute text-2xl">{progress}%</span>
-    </div>
   );
 };
